tp3/requirements/base.txt
fastapi
uvicorn
tortoise-orm
aerich
asyncpg
psycopg2-binary
fastapi
uvicorn
tortoise-orm
aerich
asyncpg
psycopg2-binary
tomli
tomlkit
